Move join method to the right place
