+++ apps/meteor/client/views/room/components/body/composer/ComposerMessage.tsx
@@ -1,5 +1,5 @@
 import type { IRoom, ISubscription } from '@rocket.chat/core-typings';
-import { useToastMessageDispatch } from '@rocket.chat/ui-contexts';
+import { useEndpoint, useToastMessageDispatch } from '@rocket.chat/ui-contexts';
 import type { ContextType, ReactElement } from 'react';
 import React, { memo, useCallback, useMemo } from 'react';
 
@@ -26,8 +26,21 @@ export type ComposerMessageProps = {
 const ComposerMessage = ({ rid, chatMessagesInstance, readOnly, onSend, ...props }: ComposerMessageProps): ReactElement => {
 	const dispatchToastMessage = useToastMessageDispatch();
 
+	const joinEndpoint = useEndpoint('POST', '/v1/channels.join');
+
 	const composerProps = useMemo(
 		() => ({
+			onJoin: async (): Promise<void> => {
+				try {
+					await joinEndpoint({
+						roomId: rid,
+					});
+				} catch (error) {
+					dispatchToastMessage({ type: 'error', message: error });
+					throw error;
+				}
+			},
+
 			onSend: async ({ value: text, tshow }: { value: string; tshow?: boolean }): Promise<void> => {
 				try {
 					await chatMessagesInstance?.flows.action.stop('typing');
